requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile
#
asgiref==3.4.1
    # via uvicorn
bcrypt==3.2.0
    # via -r requirements.in
certifi==2021.5.30
    # via requests
cffi==1.14.6
    # via bcrypt
charset-normalizer==2.0.3
    # via requests
click==8.0.1
    # via
    #   pip-tools
    #   uvicorn
colorama==0.4.4
    # via click
dnspython==2.1.0
    # via email-validator
ecdsa==0.17.0
    # via python-jose
email-validator==1.1.3
    # via pydantic
facebook-sdk==3.1.0
    # via -r requirements.in
fastapi==0.68.0
    # via -r requirements.in
h11==0.12.0
    # via uvicorn
idna==3.2
    # via
    #   email-validator
    #   requests
motor==2.5.0
    # via -r requirements.in
passlib==1.7.4
    # via -r requirements.in
pep517==0.11.0
    # via pip-tools
pip-tools==6.2.0
    # via -r requirements.in
pyasn1==0.4.8
    # via
    #   python-jose
    #   rsa
pycparser==2.20
    # via cffi
pydantic[email]==1.8.2
    # via
    #   -r requirements.in
    #   fastapi
pymongo==3.12.0
    # via
    #   -r requirements.in
    #   motor
python-decouple==3.4
    # via -r requirements.in
python-jose==3.3.0
    # via -r requirements.in
requests==2.26.0
    # via facebook-sdk
rsa==4.7.2
    # via python-jose
six==1.16.0
    # via
    #   bcrypt
    #   ecdsa
starlette==0.14.2
    # via fastapi
tomli==1.2.0
    # via pep517
typing-extensions==3.10.0.0
    # via pydantic
urllib3==1.26.6
    # via requests
uvicorn==0.14.0
    # via -r requirements.in
wheel==0.36.2
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
